Expand tab switching functionality
Now, switching tabs also saves the result label as well as the table.
When you switch to a different tab, it will use the saved data for result label and table.
Currently, badge, calculate, result label, table all work with tab switching.
Only other thing to change is the weights as quick play only has 3, not 4.