Begin new data gathering
With the introduction of season 9, awarded point  values have changed.
Quick play is no longer it's on separate rank, so that tab will be deleted in the future.
Additoinally, quick play now awards points for world tour, but a different number than the previous awarded amount for quick play badge progress. Currently I believe it is 4 for a win and 2 for a loss.
Calculator will be getting a large overhaul to include more gamemodes and updated point values for world tour.
Everything will now be under the world tour badges. No matter the game mode you play, you will gain points for world tour.